[WIP] Add config generator support
It still needs code documentation, proper configuration for the required
keys (currently just checking it works) and extra options

Improved processor configuration and messages

Better parameter handling to configure the processors

Add basic support to overwrite keys

Add basic logging to a file

Remove commented code

Modify template to ask for data

Add support to use overwrite keys from a file

The keys will be fetched first from the file (if any) and then from the
app arguments, being overwritten if needed. The resulting dictionary
will be used in the OverwriterProcessor

Add code documentation and cleanup a bit

Changes for compatibility with python3

Checked with python 2.7 and 3.4. Should work with both versions

Add the oc_webdav_endpoint to the required changes

Add a default password for the user accounts so tests won't fail